# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Axum Production API",
    "description": "Production-ready Axum API scaffold: session-backed auth, M2M API keys, RBAC, apps, organizations, owned storage with signed URLs, rate limiting, idempotency, metrics, cross-replica realtime, notifications, and an append-only audit log",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/files/org/{org_id}": {
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "List the files belonging to an organization the caller is a member of.",
        "operationId": "list_org_files",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "description": "Organization ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-indexed, default: 1)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Items per page (default: 20, max: 100)",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64",
              "default": 20,
              "maximum": 100,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Files owned by the organization",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Vec_UploadResponse"
                }
              }
            }
          },
          "404": {
            "description": "Organization not found or caller is not a member",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/files/presigned-url": {
      "post": {
        "tags": [
          "Storage"
        ],
        "operationId": "create_presigned_url",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignedUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Signed, expiring upload grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PresignedUploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unacceptable content type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/files/upload": {
      "post": {
        "tags": [
          "Storage"
        ],
        "operationId": "upload_file",
        "responses": {
          "201": {
            "description": "File uploaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_UploadResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing or unacceptable file",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "File exceeds the size limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/files/{id}": {
      "get": {
        "tags": [
          "Storage"
        ],
        "summary": "Download a file.",
        "description": "Access requires one of: the file being public, a bearer token for its owner\nor an administrator, or a valid unexpired signature. Callers who fail all\nthree get 404 rather than 403, so this cannot be used to probe for file IDs.",
        "operationId": "get_file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "expires",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "signature",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "key",
            "in": "query",
            "description": "Reserved key, present on presigned upload URLs.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File contents",
            "content": {
              "application/octet-stream": {}
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Storage"
        ],
        "operationId": "delete_file",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "File deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_String"
                }
              }
            }
          },
          "404": {
            "description": "File not found or not yours",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
    "/api/v1/files/{id}/signed-url": {
      "get": {
        "tags": [
          "Storage"
        ],
        "operationId": "create_signed_download_url",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "File ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signed, expiring download URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SignedUrlResponse"
                }
              }
            }
          },
          "404": {
            "description": "File not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiErrorPayload": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {},
          "message": {
            "type": "string"
          }
        }
      },
      "ApiErrorResponse": {
        "type": "object",
        "description": "Standardized Error Envelope",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorPayload"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponse_PresignedUploadResponse": {
        "type": "object",
        "description": "Standardized Success Envelope",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "description": "A genuinely signed upload grant. The previous implementation returned a static\npath with no signature and an advertised expiry that nothing enforced.",
            "required": [
              "storage_key",
              "upload_url",
              "file_url",
              "expires_at",
              "expires_in_seconds",
              "max_bytes"
            ],
            "properties": {
              "expires_at": {
                "type": "string",
                "format": "date-time"
              },
              "expires_in_seconds": {
                "type": "integer",
                "format": "int64"
              },
              "file_url": {
                "type": "string"
              },
              "max_bytes": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "storage_key": {
                "type": "string"
              },
              "upload_url": {
                "type": "string",
                "description": "Absolute URL carrying the signature and expiry as query parameters."
              }
            }
          },
          "meta": {},
          "success": {
            "type": "boolean"
          }
        }
      },
      "ApiResponse_SignedUrlResponse": {
        "type": "object",
        "description": "Standardized Success Envelope",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "url",
              "expires_at",
              "expires_in_seconds"
            ],
            "properties": {
              "expires_at": {
                "type": "string",
                "format": "date-time"
              },
              "expires_in_seconds": {
                "type": "integer",
                "format": "int64"
              },
              "url": {
                "type": "string"
              }
            }
          },
          "meta": {},
          "success": {
            "type": "boolean"
          }
        }
      },
      "ApiResponse_String": {
        "type": "object",
        "description": "Standardized Success Envelope",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "type": "string"
          },
          "meta": {},
          "success": {
            "type": "boolean"
          }
        }
      },
      "ApiResponse_UploadResponse": {
        "type": "object",
        "description": "Standardized Success Envelope",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "storage_key",
              "original_name",
              "size_bytes",
              "mime_type",
              "visibility",
              "url"
            ],
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "mime_type": {
                "type": "string"
              },
              "org_id": {
                "type": [
                  "string",
                  "null"
                ],
                "format": "uuid"
              },
              "original_name": {
                "type": "string"
              },
              "size_bytes": {
                "type": "integer",
                "format": "int64"
              },
              "storage_key": {
                "type": "string"
              },
              "url": {
                "type": "string",
                "description": "Relative download path. For a private file this needs a bearer token or a\nsigned URL obtained from `/files/{id}/signed-url`."
              },
              "visibility": {
                "type": "string"
              }
            }
          },
          "meta": {},
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponse_Vec_UploadResponse": {
        "type": "object",
        "description": "Standardized Success Envelope",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "storage_key",
                "original_name",
                "size_bytes",
                "mime_type",
                "visibility",
                "url"
              ],
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "mime_type": {
                  "type": "string"
                },
                "org_id": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid"
                },
                "original_name": {
                  "type": "string"
                },
                "size_bytes": {
                  "type": "integer",
                  "format": "int64"
                },
                "storage_key": {
                  "type": "string"
                },
                "url": {
                  "type": "string",
                  "description": "Relative download path. For a private file this needs a bearer token or a\nsigned URL obtained from `/files/{id}/signed-url`."
                },
                "visibility": {
                  "type": "string"
                }
              }
            }
          },
          "meta": {},
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "PresignedUploadRequest": {
        "type": "object",
        "required": [
          "filename",
          "content_type"
        ],
        "properties": {
          "content_type": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "size_bytes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "visibility": {
            "type": [
              "string",
              "null"
            ],
            "description": "`private` (default) or `public`."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}